+++ Backend/routes/Report/questions.js
@@ -3,14 +3,21 @@ const Question = require('../../models/Report/Question');
 const Report = require('../../models/Report/Report');
 const fetchuser = require('../../middleware/fetchuser');
 const UserResponse = require('../../models/Report/UserResponse');
+const User = require('../../models/User');
 const router = express.Router();
 
 // POST route to add a new question
-router.post('/add/:id', async (req, res) => {
+router.post('/add/:id', fetchuser, async (req, res) => {
   try {
     const reportId = req.params.id;
 
-    const { text, type, options, index, maxScore,scenarioId } = req.body;
+    const userAdmin = await User.findById(req.user.id);
+ 
+    if (userAdmin.role != process.env.WT) {
+      return res.status(403).send({ error: "Bad Request" });
+    }
+
+    const { text, type, options, index, maxScore, scenarioId } = req.body;
 
     // Verify the report type exists
     const reportExists = await Report.findById(reportId);
@@ -46,11 +53,16 @@ router.post('/add/:id', async (req, res) => {
 });
 
 // PUT route to update a question
-router.put('/edit/:id', async (req, res) => {
+router.put('/edit/:id', fetchuser, async (req, res) => {
   const { id } = req.params; // Get the question ID from the URL parameters
   const { text, type, options, index, maxScore, scenarioId } = req.body; // Get the updated data from the request body
 
   try {
+    const userAdmin = await User.findById(req.user.id);
+ 
+    if (userAdmin.role != process.env.WT) {
+      return res.status(403).send({ error: "Bad Request" });
+    }
 
     // Validate scenarioId
     if (!scenarioId) {
@@ -83,28 +95,39 @@ router.put('/edit/:id', async (req, res) => {
 });
 
 // DELETE route to remove a question
-router.delete('/delete/:id', async (req, res) => {
+router.delete('/delete/:id', fetchuser, async (req, res) => {
   const { id } = req.params; // Get the question ID from the URL parameters
 
   try {
-      // Find the question by ID and delete it
-      const deletedQuestion = await Question.findByIdAndDelete(id);
+    const userAdmin = await User.findById(req.user.id);
+ 
+    if (userAdmin.role != process.env.WT) {
+      return res.status(403).send({ error: "Bad Request" });
+    }
+    // Find the question by ID and delete it
+    const deletedQuestion = await Question.findByIdAndDelete(id);
 
-      if (!deletedQuestion) {
-          return res.status(404).json({ message: 'Question not found' });
-      }
+    if (!deletedQuestion) {
+      return res.status(404).json({ message: 'Question not found' });
+    }
 
-      res.json({ message: 'Question deleted successfully' });
+    res.json({ message: 'Question deleted successfully' });
   } catch (error) {
-      console.error('Error deleting question:', error);
-      res.status(500).json({ message: 'Failed to delete question' });
+    console.error('Error deleting question:', error);
+    res.status(500).json({ message: 'Failed to delete question' });
   }
 });
 
 
 // Get questions for a specific report for admins only
-router.get('/for/:reportId/', async (req, res) => {
+router.get('/for/:reportId/', fetchuser, async (req, res) => {
   try {
+    const userAdmin = await User.findById(req.user.id);
+ 
+    if (userAdmin.role != process.env.WT) {
+      return res.status(403).send({ error: "Bad Request" });
+    }
+
     // Fetch the questions and populate scenarioId
     const questions = await Question.find({ report: req.params.reportId })
       .sort({ index: 1 })
@@ -142,8 +165,8 @@ router.get('/for/:reportId/:scenarioId', fetchuser, async (req, res) => {
     const existingResponse = await UserResponse.findOne({ reportId, scenarioId, userId });
 
     if (existingResponse) {
-      return res.status(403).json({ 
-        message: 'You have already submitted this report. Editing is not allowed.' 
+      return res.status(403).json({
+        message: 'You have already submitted this report. Editing is not allowed.'
       });
     }
 
@@ -157,9 +180,9 @@ router.get('/for/:reportId/:scenarioId', fetchuser, async (req, res) => {
     res.status(200).json(questions);
   } catch (error) {
     console.error('Error fetching questions:', error);
-    res.status(500).json({ 
-      message: 'An error occurred while fetching questions.', 
-      error: error.message 
+    res.status(500).json({
+      message: 'An error occurred while fetching questions.',
+      error: error.message
     });
   }
 });
